+++ src/main/java/de/madetv1280/craftD3K/commands/CraftCommand.java
@@ -3,6 +3,9 @@ package de.madetv1280.craftD3K.commands;
 import de.madetv1280.craftD3K.constants.keys.Config;
 import de.madetv1280.craftD3K.constants.keys.messages.Msg;
 import de.madetv1280.craftD3K.constants.Permission;
+import de.madetv1280.craftD3K.managers.ConfigManager;
+import de.madetv1280.craftD3K.managers.CooldownManager;
+import de.madetv1280.craftD3K.util.PluginLogger;
 import org.bukkit.Bukkit;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandExecutor;
@@ -13,17 +16,31 @@ import org.bukkit.inventory.Inventory;
 import org.bukkit.plugin.java.JavaPlugin;
 import org.jetbrains.annotations.NotNull;
 
+import java.util.UUID;
+import java.util.concurrent.TimeUnit;
+
 import static de.madetv1280.craftD3K.managers.MessageManager.getMessage;
 
 public class CraftCommand implements CommandExecutor {
     private final JavaPlugin plugin;
+    private final CooldownManager cooldownManager;
 
     public CraftCommand(JavaPlugin plugin) {
         this.plugin = plugin;
+        this.cooldownManager = new CooldownManager();
     }
 
     @Override
     public boolean onCommand(@NotNull CommandSender sender, @NotNull Command command, @NotNull String s, @NotNull String @NotNull [] strings) {
+
+        // Get all config values
+        boolean bypassCooldownAsOp = ConfigManager.getBoolean(Config.BYPASS_COOLDOWN_AS_OP, false);
+        boolean showCooldownMsg = ConfigManager.getBoolean(Config.COOLDOWN_MSG, false);
+        boolean showOpenMenuMsg = ConfigManager.getBoolean(Config.SHOW_OPEN_MENU_MSG, false);
+
+        int cooldown = ConfigManager.getInt(Config.COOLDOWN, 0);
+        long cooldownInMillis = TimeUnit.SECONDS.toMillis(cooldown);
+
 
         // Check if sender is a player
         if (!(sender instanceof Player player)) {
@@ -38,14 +55,27 @@ public class CraftCommand implements CommandExecutor {
         }
 
 
-        // Cooldown logic
-        if (!sender.hasPermission(Permission.BYPASS_COOLDOWN)) {
-            // Cooldown logic
+        boolean hasBypassPermission = sender.hasPermission(Permission.BYPASS_COOLDOWN);
+        boolean isOpWithBypassEnabled = sender.isOp()
+                && bypassCooldownAsOp;
+
+        UUID uuid = player.getUniqueId();
+
+        // Check for cooldown
+        if (!hasBypassPermission && !isOpWithBypassEnabled) {
+
+            if (cooldownManager.isOnCooldown(uuid, cooldownInMillis)) {
+                if (showCooldownMsg) {
+                    long remaining = (cooldownManager.getRemainingSeconds(uuid, cooldownInMillis));
+                    sender.sendMessage(getMessage(Msg.COOLDOWN).replace("%s", String.valueOf(remaining)));
+                }
+                return true;
+            }
+            cooldownManager.setCooldown(uuid);
         }
 
-
-        // Check if opening-msg is needed
-        if (plugin.getConfig().getBoolean(Config.SHOW_OPEN_MENU_MSG)) {
+        // Check if msg is needed
+        if (showOpenMenuMsg) {
             sender.sendMessage(getMessage(Msg.OPENING_MENU));
         }