+++ src/main/java/de/madetv1280/craftD3K/commands/CraftReloadCommand.java
@@ -31,6 +31,8 @@ public class CraftReloadCommand implements CommandExecutor {
         Bukkit.getPluginManager().disablePlugin(plugin);
         Bukkit.getPluginManager().enablePlugin(plugin);
 
+        plugin.reloadConfig();
+
         sender.sendMessage(getMessage(Msg.PLUGIN_RELODED));
         return true;
     }