+++ src/main/java/de/madetv1280/craftD3K/constants/keys/Config.java
@@ -9,4 +9,8 @@ public class Config {
     public static final String SHOW_OPEN_MENU_MSG = "opening-menu-msg";
     public static final String DEBUG = "debug";
 
+    public static final String COOLDOWN = "cooldown";
+    public static final String COOLDOWN_MSG = "cooldown-msg";
+    public static final String BYPASS_COOLDOWN_AS_OP = "bypass-cooldown-as-op";
+
 }