+++ src/main/java/de/madetv1280/craftD3K/constants/keys/messages/InternalMsg.java
@@ -11,8 +11,9 @@ public class InternalMsg {
     public static final String RUNNING_OLDER_VERSION = "running-older-version";
     public static final String RUNNING_LATEST_VERSION = "running-lastest-version";
     public static final String FAILED_CHECK_UPDATE_UNKNOW_VERSION_FORMAT = "failed-check-update-unknown-version-format";
-    public static final String FAILED_CHECK_UPDATE_NETWORK_CONNTECTION = "failed-check-update-network-connection";
+    public static final String FAILED_CHECK_UPDATE_NETWORK_CONNECTION = "failed-check-update-network-connection";
     public static final String DISABLE_PLUGIN = "disable-plugin";
     public static final String DEBUG_ENABLED = "debug-enabled";
-
+    public static final String INCORRECT_DATATYPE = "incorrect-datatype";
+    public static final String COMMAND_MISSING = "command-missing";
 }